pyproject.toml
[project]
name = "file-watcher"
readme = "README.md"
version = "0.0.1"
dependencies = [
    "memphis-py==1.0.1",
    "psycopg2==2.9.6",
    "SQLAlchemy==2.0.13"
]


[project.urls]
"Repository" = "https://github.com/interactivereduction/file-watcher"

[project.scripts]
file_watcher = "file_watcher.main:main"

[project.optional-dependencies]
formatting = [
    "black==23.3.0"
]

test = [
    "pytest==7.3.2",
    "pytest-cov==4.0.0",
    "pytest-asyncio==0.21.0"
]

code-inspection = [
    "pylint==2.17.4",
    "mypy==1.2.0",
    "file_watcher[test]"
]

dev = [
    "file_watcher[code-inspection]"
]


[tool.setuptools]
packages = ["file_watcher"]
